Fix bug in pvec `pop` resulting in empty tail
When calling `pop` on a persistent vector with, eg., 33 elements, we have a `tail` of length 1.  After this `pop` call, there was a bug where the resulting pvec had a tail of length `0` rather than a tail of length `32`.  This causes errors, for instance when calling `iterate`, which expects a nonempty tail.  This PR fixes the behavior by adding a branch in `pop` function for when the tail length is currently `1` (in which case we make the new tail be `pop(trie)`).